Add rc.d/buzzkill script for FreeBSD
Start buzzkill.js using daemon(8) to ensure it restarts even after a
crash, as the unpriviledged user 'nobody' , and log everything to
syslog.

This rc.d script is relocatable and expects to find buzzkill.js in its
parent directory. In other words, it should be run directly from the
git clone, not copied to /usr/local/etc/rc.d/