rc.d/buzzkill: add buzzkill_dir rcvar
rc(8) on FreeBSD uses run_rc_script from rc.subr, which sources the rc
script instead of executing it. This means the `dirname $0` trick to
determine the location of the buzzkill.js script doesn't work (since $0
is /etc/rc, not the rc.d/buzzkill script).

Add a buzzkill_dir rcvar, which defaults to ~ec2-user/buzzkill, and use
this variable to locate buzzkill.js